+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * Keep command payload
   * Fix channels info for scene status commands (`0x0003`, `0xEFFF`)
   * Fix parsing of "Response Read Sensors Status" (`0x1605`)
+  * Add Panel buttons control commands (`0xE012`, `0xE14E`)
 
 **BREAKING:**